Extracted super_selectable_text package (Resolves #474)
