+++ attributed_text/lib/attributed_text.dart
@@ -2,4 +2,5 @@ export 'src/attributed_text.dart';
 export 'src/attributed_spans.dart';
 export 'src/attribution.dart';
 export 'src/test_tools.dart';
+export 'src/text_tools.dart';
 export 'src/span_range.dart';
