+++ super_editor/lib/src/core/document_selection.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:super_editor/src/default_editor/text.dart';
-import 'package:super_editor/src/infrastructure/super_selectable_text.dart';
 
 import 'document.dart';
