+++ super_editor/lib/src/default_editor/box_component.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:super_editor/src/core/document.dart';
 import 'package:super_editor/src/default_editor/selection_upstream_downstream.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
-import 'package:super_editor/src/infrastructure/caret.dart';
+import 'package:super_selectable_text/super_selectable_text.dart';
 
 import '../core/document_layout.dart';
