+++ super_editor/lib/src/infrastructure/super_textfield/ios/ios_textfield.dart
@@ -10,9 +10,9 @@ import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/h
 import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/text_scrollview.dart';
 import 'package:super_editor/src/infrastructure/super_textfield/input_method_engine/_ime_text_editing_controller.dart';
 import 'package:super_editor/src/infrastructure/super_textfield/ios/_editing_controls.dart';
+import 'package:super_selectable_text/super_selectable_text.dart';
 
 import '../../platforms/ios/toolbar.dart';
-import '../../super_selectable_text.dart';
 import '_caret.dart';
 import '_floating_cursor.dart';
 import '_user_interaction.dart';
