+++ super_editor/lib/super_editor.dart
@@ -1,13 +1,12 @@
 library super_editor;
 
 export 'package:attributed_text/attributed_text.dart';
+export 'package:super_selectable_text/super_selectable_text.dart';
 export 'src/infrastructure/attributed_text_styles.dart';
 export 'src/infrastructure/super_textfield/super_textfield.dart';
-export 'src/infrastructure/super_selectable_text.dart';
 export 'src/infrastructure/_listenable_builder.dart';
 export 'src/infrastructure/_logging.dart';
 export 'src/infrastructure/multi_tap_gesture.dart';
-export 'src/infrastructure/text_layout.dart';
 export 'src/core/document.dart';
 export 'src/core/document_composer.dart';
 export 'src/core/document_editor.dart';
