+++ super_editor/test/src/infrastructure/super_textfield/desktop/super_desktop_textfield_keyboard_test.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart' hide SelectableText;
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:super_editor/src/infrastructure/platform_detector.dart';
-import 'package:super_editor/src/infrastructure/super_selectable_text.dart';
 import 'package:super_editor/src/infrastructure/super_textfield/super_textfield.dart';
+import 'package:super_selectable_text/super_selectable_text.dart';
 
 import '../../../_text_entry_test_tools.dart';
 import '../../_platform_test_tools.dart';
