+++ super_selectable_text/README.md
@@ -1,0 +1,3 @@
+# Super Selectable Text
+Configurable, composable, extensible text display for Flutter.
+
