+++ super_selectable_text/lib/src/text_layout.dart
@@ -3,8 +3,8 @@ import 'package:flutter/widgets.dart';
 
 /// Contract to interrogate the layout of a blob of text.
 abstract class TextLayout {
-  /// Returns [true] if a text character overlaps the given [localOffset],
-  /// or [false] otherwise.
+  /// Returns `true` if a text character overlaps the given [localOffset],
+  /// or `false` otherwise.
   bool isTextAtOffset(Offset localOffset);
 
   /// Returns the [TextPosition] that overlaps the given [localOffset].
@@ -59,22 +59,22 @@ abstract class TextLayout {
   /// in the line of text that contains the given [textPosition].
   TextPosition getPositionAtEndOfLine(TextPosition textPosition);
 
-  /// Returns the `TextPosition` in the first line within this
-  /// `TextLayout` that is closest to the given `x`-value, or
-  /// -1 if the text is not laid out yet.
+  /// Returns the [TextPosition] in the first line within this
+  /// [TextLayout] that is closest to the given `x`-value, or
+  /// `-1` if the text is not laid out yet.
   TextPosition getPositionInFirstLineAtX(double x);
 
-  /// Returns the `TextPosition` in the last line within this
-  /// `TextLayout` that is closest to the given `x`-value, or
-  /// -1 if the text is not laid out yet.
+  /// Returns the [TextPosition] in the last line within this
+  /// [TextLayout] that is closest to the given `x`-value, or
+  /// `-1` if the text is not laid out yet.
   TextPosition getPositionInLastLineAtX(double x);
 
-  /// Returns the `TextSelection` that corresponds to a selection
-  /// rectangle formed by the span from `baseOffset` to `extentOffset`, or
-  /// a collapsed selection at -1 if the text is not laid out yet.
+  /// Returns the [TextSelection] that corresponds to a selection
+  /// rectangle formed by the span from [baseOffset] to [extentOffset], or
+  /// a collapsed selection at `-1` if the text is not laid out yet.
   ///
-  /// The `baseOffset` determines where the selection begins. The
-  /// `extentOffset` determines where the selection ends.
+  /// The [baseOffset] determines where the selection begins. The
+  /// [extentOffset] determines where the selection ends.
   TextSelection getSelectionInRect(Offset baseOffset, Offset extentOffset);
 
   /// Returns a [TextSelection] that surrounds the given [startingPosition] and expands
@@ -82,8 +82,12 @@ abstract class TextLayout {
   TextSelection expandSelection(TextPosition startingPosition, TextExpansion expansion, TextAffinity affinity);
 }
 
+/// Function that expands from a given [startingPosition] to an expanded
+/// [TextSelection], based on some expansion decision behavior.
 typedef TextExpansion = TextSelection Function(String text, TextPosition startingPosition, TextAffinity affinity);
 
+/// [TextExpansion] function that expands from [startingPosition] in both directions
+/// to select a single paragraph.
 TextSelection paragraphExpansionFilter(String text, TextPosition startingPosition, TextAffinity affinity) {
   // If the given position falls directly on a newline then return
   // just the newline character as the paragraph selection.
@@ -110,75 +114,4 @@ TextSelection paragraphExpansionFilter(String text, TextPosition startingPositio
           baseOffset: end,
           extentOffset: start,
         );
-}
-
-/// Returns the code point index for the code point that ends the visual
-/// character that begins at [startingCodePointIndex].
-///
-/// A single visual character might be comprised of multiple code points.
-/// Each code point occupies a slot within a [String], which means that
-/// an index into a [String] might refer to a piece of a single visual
-/// character.
-///
-/// [startingCodePointIndex] is the traditional [String] index for the
-/// leading code point of a visual character.
-///
-/// This function starts at the given [startingCodePointIndex] and walks
-/// towards the end of [text] until it has accumulated an entire
-/// visual character. The [String] index of the final code point for
-/// the given character is returned.
-int getCharacterEndBounds(String text, int startingCodePointIndex) {
-  // This implementation was copied and adapted from text_editing_action_target
-  // in the Flutter repo.
-  assert(startingCodePointIndex >= 0 && startingCodePointIndex <= text.length);
-
-  if (startingCodePointIndex == text.length) {
-    return text.length;
-  }
-
-  final CharacterRange range = CharacterRange.at(text, 0, startingCodePointIndex);
-  // If index is not on a character boundary, return the next character
-  // boundary.
-  if (range.current.length != startingCodePointIndex) {
-    return range.current.length;
-  }
-
-  range.expandNext();
-  return range.current.length;
-}
-
-/// Returns the code point index for the code point that begins the visual
-/// character that ends at [endingCodePointIndex].
-///
-/// A single visual character might be comprised of multiple code points.
-/// Each code point occupies a slot within a [String], which means that
-/// an index into a [String] might refer to a piece of a single visual
-/// character.
-///
-/// [endingCodePointIndex] is the traditional [String] index for the
-/// trailing code point of a visual character.
-///
-/// This function starts at the given [endingCodePointIndex] and walks
-/// towards the beginning of [text] until it has accumulated an entire
-/// visual character. The [String] index of the initial code point for
-/// the given character is returned.
-int getCharacterStartBounds(String text, int endingCodePointIndex) {
-  // This implementation was copied and adapted from text_editing_action_target
-  // in the Flutter repo.
-  assert(endingCodePointIndex >= 0 && endingCodePointIndex <= text.length);
-
-  if (endingCodePointIndex == 0) {
-    return 0;
-  }
-
-  final CharacterRange range = CharacterRange.at(text, 0, endingCodePointIndex);
-  // If index is not on a character boundary, return the previous character
-  // boundary.
-  if (range.current.length != endingCodePointIndex) {
-    range.dropLast();
-    return range.current.length;
-  }
-
-  range.dropLast();
-  return range.current.length;
 }
