+++ super_selectable_text/lib/super_selectable_text.dart
@@ -1,0 +1,3 @@
+export 'src/caret.dart';
+export 'src/super_selectable_text.dart';
+export 'src/text_layout.dart';
