+++ super_selectable_text/pubspec.yaml
@@ -1,0 +1,25 @@
+name: super_selectable_text
+description: Configurable, composable, extensible text display for Flutter.
+version: 0.1.0
+homepage: https://github.com/superlistapp/super_editor
+
+environment:
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.17.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+
+  attributed_text:
+    path: ../attributed_text
+  flutter_lints: ^1.0.0
+  logging: ^1.0.1
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  golden_toolkit: ^0.11.0
+
+flutter:
+  # no Flutter configuration
