+++ super_selectable_text/test/super_selectable_text_test.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  group("Super selectable text", () {
+    // TODO: define some standard use-cases to lock down the API
+    // and show standard usage.
+  });
+}
